src/requirements.txt
en-core-web-sm==3.0.0
luigi==3.0.2
matplotlib==3.1.1
pandas==1.0.3
pytesseract==0.3.3
Pillow==6.2.1
pymongo==3.11.3
spacy==3.0.0
Wand==0.6.5
scikit-learn==0.22.1